+++ R/run_app.R
@@ -21,9 +21,7 @@ assist_data_addin <- function() {
 #' Low Code Assistant - Transformations
 #'
 #' Run the data transformation LCA wizard.
-#' @param data_name A data frame to use. Can either be a data frame object, the name of
-#' a data frame that exists in your global environment, or a reactive string with the name
-#' of a data frame.
+#' @param data_name The name of a data frame in the global environment. Can be reactive or static.
 #' @export
 assist_transform <- function(data_name = NULL) {
   app <- shiny::shinyApp(
@@ -58,21 +56,36 @@ assist_transform_addin <- function() {
 #' Low Code Assistant - Visualizations
 #'
 #' Run the data visualizations LCA wizard.
+#' @param data_name The name of a data frame in the global environment. Can be reactive or static.
 #' @export
-assist_viz <- function() {
+assist_viz <- function(data_name = NULL) {
   app <- shiny::shinyApp(
     ui = shiny::fluidPage(page_viz_ui("app")),
-    server = function(input, output, session) page_viz_server("app")
+    server = function(input, output, session) page_viz_server("app", data_name_in = data_name)
   )
   shiny::runGadget(app, viewer = shiny::dialogViewer("Domino R Assistant", 1100, 800))
 }
 
 #' Low Code Assistant - Visualizations
 #'
-#' Run the data visualizations LCA wizard.
+#' Run the data visualizations LCA wizard. If a variable holding a data frame is selected
+#' when invoking the addin in RStudio, that data frame will be used.
 #' @export
 assist_viz_addin <- function() {
-  assist_viz()
+  data_name <- tryCatch({
+    context <- rstudioapi::getActiveDocumentContext()
+    name <- context$selection[[1]]$text
+
+    if (nzchar(name)) {
+      name
+    } else {
+      NULL
+    }
+  }, error = function(err) {
+    NULL
+  })
+
+  assist_viz(data_name = data_name)
 }
 
 #' Low Code Assistant - Snippets